Add Asyncify backend for emscripten (#1)
Unlike the 'emscripten' backend, this one doesn't use the emscripten
coroutines API (which is currently broken) and instead implements a
minimal context-switching "fiber" API as a JS library on top of
Asyncify. See emscripten-core/emscripten#8979 for more details.

`_koishi_impl_fiber_swap` must be manually added to `ASYNCIFY_IMPORTS`
by the user of the library, because emcc does not provide a way to
append to the imports array without completely overwriting it yet.