Add new backend "ucontext_sjlj"
Uses ucontext APIs to set up contexts and sigjmp/longjmp to switch
between them. Setup is more expensive than in the regular ucontext
backend, but switching is much faster.

This may not work on some systems where these APIs are available, so
this backend won't be picked automatically. Consider switching to it
manually if impl=auto picks ucontext.