api: define new KOISHI_IDLE state
This is the state of coroutines that have resumed another coroutine and
are waiting for it to yield. Idle coroutines can not be resumed; they
must yield in order to become Suspended (and thus resumable) again.

Previously such coroutines were considered suspended, and could be
resumed, but this was never supported. Doing so would mess up the
call-chain and possibly lead to assertion failures. This is explicitly
prohibited now, and attempting to resume an idle coroutine will trip an
early assertion in koishi_resume.